ai-database-agent/requirements.txt
streamlit
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-cloud-aiplatform
langchain-anthropic
langchain-google-vertexai
langchain
langchain-community
uuid